Refactor terminal with command registry and autocomplete
